refactor(spend_book): factor out the spendbook from mint.rs
