+++ src/dbc.rs
@@ -61,12 +61,7 @@ impl Dbc {
 
             verifier
                 .verify(&self.transaction.hash(), mint_key, mint_sig)
-                .map_err(|e| {
-                    Error::Signing(format!(
-                        "Failed to verify mint signature: {:?}",
-                        e.to_string()
-                    ))
-                })?;
+                .map_err(|e| Error::Signing(e.to_string()))?;
         }
         if self.transaction.inputs.is_empty() {
             Err(Error::TransactionMustHaveAnInput)
